+++ src/main/java/com/robertx22/age_of_exile/database/registrators/Stats.java
@@ -10,6 +10,7 @@ import com.robertx22.age_of_exile.database.data.stats.types.generated.*;
 import com.robertx22.age_of_exile.database.data.stats.types.loot.TreasureQuality;
 import com.robertx22.age_of_exile.database.data.stats.types.loot.TreasureQuantity;
 import com.robertx22.age_of_exile.database.data.stats.types.misc.BonusExp;
+import com.robertx22.age_of_exile.database.data.stats.types.misc.BonusFavor;
 import com.robertx22.age_of_exile.database.data.stats.types.misc.ChangeDmgElementStat;
 import com.robertx22.age_of_exile.database.data.stats.types.misc.ExtraMobDropsStat;
 import com.robertx22.age_of_exile.database.data.stats.types.offense.*;
@@ -122,6 +123,7 @@ public class Stats implements ISlashRegistryInit {
 
                     add(ExtraMobDropsStat.getInstance());
                     add(BonusExp.getInstance());
+                    add(BonusFavor.getInstance());
 
                     add(new UnknownStat());