+++ src/main/java/com/robertx22/age_of_exile/saveclasses/player_skills/PlayerSkillEnum.java
@@ -19,6 +19,7 @@ public enum PlayerSkillEnum {
     INSCRIBING("inscribing", Words.Inscribing, Formatting.AQUA, null, () -> Items.INK_SAC),
     COOKING("cooking", Words.Cooking, Formatting.RED, null, () -> Items.BREAD),
     FISHING("fishing", Words.Fishing, Formatting.BLUE, EnchantmentTarget.FISHING_ROD, () -> Items.FISHING_ROD),
+    EXPLORATION("exploration", Words.Exploration, Formatting.GOLD, EnchantmentTarget.ARMOR_HEAD, () -> Items.CHEST),
     NONE("none", Words.Elemental_Attack_Damage, Formatting.BLACK, null, () -> Items.BEDROCK);
 
     public String id;