+++ src/main/java/com/robertx22/age_of_exile/uncommon/localization/Words.java
@@ -6,6 +6,7 @@ import com.robertx22.age_of_exile.uncommon.interfaces.IAutoLocName;
 import java.util.Locale;
 
 public enum Words implements IAutoLocName {
+    Exploration("Exploration"),
     Fishing("Fishing"),
     Inscribing("Inscribing"),
     Cooking("Cooking"),