+++ src/main/java/com/robertx22/age_of_exile/uncommon/utilityclasses/LevelUtils.java
@@ -60,7 +60,7 @@ public class LevelUtils {
             lvl = determineLevelPerDistanceFromSpawn(sw, pos, dimConfig);
         }
 
-        lvl = MathHelper.clamp(dimConfig.min_lvl + lvl, dimConfig.min_lvl, dimConfig.max_lvl);
+        lvl = MathHelper.clamp(lvl, dimConfig.min_lvl, dimConfig.max_lvl);
 
         lvl = MathHelper.clamp(lvl, 1, ModConfig.get().Server.MAX_LEVEL);
 
@@ -98,7 +98,7 @@ public class LevelUtils {
 
         int lvl = 1;
 
-        lvl = (int) (1 + (distance / (config.mob_lvl_per_distance)));
+        lvl = (int) (config.min_lvl + (distance / (config.mob_lvl_per_distance)));
 
         return MathHelper.clamp(lvl, config.min_lvl, config.max_lvl);